add Produto model and payment

+++ database/migrations/2023_05_27_034618_create_compradors_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->date('birth_date');
             $table->string('state');
             $table->string('city');
-            $table->integer('credits');
+            $table->float('credits');
             $table->timestamps();
         });
     }
